codigo organizado
