+++ Projeto/Includes/hitable_list.cpp
@@ -1,0 +1,15 @@
+#include "./hitable_list.h"
+
+bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const {
+    hit_record temp_rec;
+    bool hit_anything = false;
+    double closet_so_far = t_max;
+    for (int i = 0; i < list_size; i++) {
+        if(list[i]->hit(r, t_min, closet_so_far, temp_rec)) {
+            hit_anything = true;
+            closet_so_far = temp_rec.t;
+            rec = temp_rec;
+        }
+    }
+    return hit_anything;
+}
