+++ Projeto/Includes/ray.cpp
@@ -1,5 +1,5 @@
 #include "ray.h"
-#include "../../External/glm/glm.hpp"
+//#include "../../External/glm/glm.hpp"
 
 // aqui estão as implementações da classe definida em ray.h
 ray::ray()
