+++ Projeto/Includes/spheres.cpp
@@ -1,0 +1,33 @@
+#include "./spheres.h"
+
+bool sphere::hit(const ray& r, float t_min, float t_max, hit_record& rec) const {
+    glm::vec3 oc = r.location() - center; // vetor que vai do centro da esfera até a origem do raio
+    float a = dot(r.direction(), r.direction());
+    float b = dot(oc, r.direction());
+    float c = dot(oc, oc) - radius*radius; // calcula o dot product de oc com ele mesmo, e subtrai o raio ao quadrado
+    //a, b e c são os coeficientes da equação de segundo grau
+    float discriminant = b*b - a*c;
+    if(discriminant > 0)
+    {
+        float temp = (-b - sqrt(b*b-a*c))/a;
+        if(temp < t_max && temp > t_min)
+        {
+            rec.t = temp;
+            rec.p = r.point_at_parameter(rec.t);
+            rec.normal = (rec.p - center) / radius;
+            rec.cor = cor;
+            return true;
+        }
+        temp = (-b + sqrt(b*b-a*c))/a;
+        if(temp < t_max && temp > t_min)
+        {
+            rec.t = temp;
+            rec.p = r.point_at_parameter(rec.t);
+            rec.normal = (rec.p - center) / radius;
+            rec.cor = cor;
+            return true;
+        }
+    }
+    
+    return false;
+}
